Login and registration

+++ Project1/WebStoreApp/WebStoreApp.Web/Controllers/LocationController.cs
@@ -3,6 +3,7 @@ using System.Web;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -22,7 +23,7 @@ namespace WebStoreApp.Web
             this._logger = logger;
         }
 
-        public async Task<IActionResult> Index(ProductModel productModel, Guid? id)
+        public async Task<IActionResult> Index(Guid? id)
         {
             if (!id.HasValue) return RedirectToAction(nameof(Index), "Locations");
 
@@ -30,6 +31,9 @@ namespace WebStoreApp.Web
             if (locationModel == null) return RedirectToAction(nameof(Index), "Locations");
 
             var productsModel = await _service.GetLocationProducts(id);
+
+            ViewData["Role"] = HttpContext.User.FindFirst(claim => claim.Type == ClaimTypes.Role)?.Value;
+
             return View(new LocationViewModel
             {
                 LocationModel = locationModel,
@@ -42,21 +46,20 @@ namespace WebStoreApp.Web
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("LocationId", "ProductName", "ProductPrice", "ProductQuantity")] ProductModel productModel, Guid? id)
         {
-            if (ModelState.IsValid && productModel.LocationId == id)
+            if (ModelState.IsValid && productModel.LocationId == id && HttpContext.User.FindFirst(claim => claim.Type == ClaimTypes.Role)?.Value == "Admin")
             {
                 await _service.CreateNewProduct(productModel);
                 return RedirectToAction(nameof(Index), "Location", new { id });
             }
 
-            ViewData["State"] = "Create";
-            return RedirectToAction(nameof(Index), "Location", new { productModel, id });
+            return RedirectToAction(nameof(Index), "Location", new { id });
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit([Bind("LocationId", "ProductId", "ProductName", "ProductPrice", "ProductQuantity")] ProductModel productModel, Guid? id)
         {
-            if (ModelState.IsValid && productModel.LocationId == id)
+            if (ModelState.IsValid && productModel.LocationId == id && HttpContext.User.FindFirst(claim => claim.Type == ClaimTypes.Role)?.Value == "Admin")
                 await _service.EditProduct(productModel);
 
             return RedirectToAction(nameof(Index), "Location", new { id });
@@ -66,7 +69,7 @@ namespace WebStoreApp.Web
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Delete([Bind("LocationId", "ProductId")] ProductModel productModel, Guid? id)
         {
-            if (productModel.LocationId == id)
+            if (productModel.LocationId == id && HttpContext.User.FindFirst(claim => claim.Type == ClaimTypes.Role)?.Value == "Admin")
                 await _service.DeleteProduct(productModel);
 
             return RedirectToAction(nameof(Index), "Location", new { id });
@@ -76,9 +79,11 @@ namespace WebStoreApp.Web
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Order(OrdersModel ordersModel, Guid? id)
         {
-            await _service.PlaceOrders(ordersModel);
+            Guid userId;
+            if (Guid.TryParse(HttpContext.User.FindFirst(claim => claim.Type == ClaimTypes.Sid)?.Value, out userId))
+                await _service.PlaceOrders(ordersModel, userId);
 
-            return RedirectToAction(nameof(Index), "Location", new { id = id });
+            return RedirectToAction(nameof(Index), "Location", new { id });
         }
     }
 }
